First working version of reviewer
reviewer_test is a sample change dict. You can run reviewer.py as module, use arrow keys to navigate, delete to discard. I do not like that we are using (window destroy) as the signal for the conman to start uploading here--it feels too implicit and unexpected, as a programmer. But I do not know of a better architecture yet.